Create create_safegraph_db.sql
This sql file will create a database and table for Safegraph's monthly_patterns data. I created it using MySQL Workbench and then forward engineered it into this script. This script is suppose to work in conjunction with load_data_into_db.sql and a csv file downloaded from Safegraph in order to create a populated table.