Added export_data.ipynb and the equivalent html
Added export_data.ipynb and the equivalent html (for uploading to Databricks). Both of these files, when used in Databricks with the monthly_patterns database table will display a table of 100,000 rows that you are able to download into a csv file on your local computer. But in a more general use (you can change some stuff), you can display and then download any data from Databricks. There is a download button below where the table is displayed.